Extracted ImageHelper and ImageController from ServiceController for reuse
